requirements.txt
# Dependencias principales del proyecto
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
python-telegram-bot[job-queue]>=20.0

# Internacionalización (detección y traducción)
langdetect>=1.0.9
deep-translator>=1.11.0

# Fanweb / API (Flask)
flask>=2.3.0

# Voz: Text-to-Speech (opcional, para audios del bot)
gtts>=2.4.0
edge-tts>=6.1.0
gradio_client>=1.3.0
huggingface_hub>=0.20

# Para automatización de navegador
# Después de instalar, ejecuta: playwright install chromium

# Tablero de Control (Dashboard) - Luz de Aurora
streamlit>=1.28.0
